backend/requirements.txt
Flask>=3.0.0
Flask-CORS>=4.0.0
Flask-SQLAlchemy>=3.0.0
PyJWT>=2.8.0
python-docx>=1.1.0
Werkzeug>=3.0.0
gunicorn>=21.0.0
